Prevent OpenSSH from reading stdin / the terminal
This appears to fix broken behaviour when doing an interactive apply: On
rare occasions, key presses were not registered by child workers. You
had to supply them twice.

OpenSSH is famous for draining all input to its stdin. For example:

    #!/bin/bash

    cat >my_hosts <<EOF
    foo.bar
    baz.bam
    yadda.yadda
    EOF

    while read host
    do
        ssh "$host" uname >"$host.out" 2>&1
    done <my_hosts

This loop has only one iteration. `ssh` read all input from stdin which
just happens to be the file `my_hosts`. So, after the first iteration,
no data is left for `read`.

Something similar happens in bw. A child worker blocks at `input()`, but
in the meantime, OpenSSH tries to read from stdin every now and then.
Sometimes, OpenSSH wins.

It's probably not *required* to use `PIPE` for `scp`. But for the sake
of consistency -- and to save us potential future trouble --, I added it
there as well.