thread safety for partial metadata access
I have not observed this being a problem in practice, but this should prevent two threads from both trying to compile metadata at the same time. Most of the time, this would only be a waste of resources, but I believe in some cases it could cause metadata to disappear.