pkg_openbsd: Support flavors
Setting "version" to select a flavor is a bit flawed because you have
to, well, select a *version*. This is not what you actually want most of
the time (at least I don't). I want to be able to select a particular
flavor and, more importantly, it shall default to the "normal" flavor if
unset.

Some examples:

    pkg_add foo
        Installs most recent version of package "foo". If there are
        multiple flavors of that package, users will be asked
        interactively to select one of them.

    pkg_add foo--
        Installs most recent version of package "foo" and selects the
        "normal" flavor.

    pkg_add foo--bar
        Installs most recent version of package "foo" and selects the
        "bar" flavor.

    pkg_add foo-1.0.0-bar
        Explicitly installs version "1.0.0" of flavor "bar".

With this patch, bw will do a "pkg_add foo--" by default. If you set a
flavor, "pkg_add foo--bar" will be used. If you set a version, the
flavor attribute will be ignored completely and "pkg_add
foo-version-bar" will be used.

I had to adjust PKGSPEC_REGEX because it was too greedy.