k8s: Ignore 'status' when comparing state
According to [0], this is something that's managed by k8s or stuff
running in it. We usually ignore this field implicitly due to
reduce_dict(), because manifests supplied by our users usually do not
include this field.

Sometimes, though, 'status' is supplied in manifests. See cert-manager's
CustomResourceDefinitions, for example. When you try to deploy those via
bw, items will always need to be fixed, because the app in k8s has
changed 'status'.

Hence, ignore 'status' in cdict() and sdict(), but still supply it in
fix().

[0]: https://book-v1.book.kubebuilder.io/basics/status_subresource.html