Fault: Hash kwargs recursively
Mutable objects are not hashable in Python. We'd like to pass those as
arguments to Faults, though. For example:

    def whatever(data):
        return json.dumps(data)

    f = Fault('foo', whatever, data={'what': 'ever', 'some': other_fault})

Mutable objects are not hashable, because, well, they're mutable. You
cannot determine a hash, because that object might change later
in-place.

Strictly speaking, this is a problem for us, too:

    def my_function(data):
        return f'{repr(data)}'

    dirty_dict = {
        'foo': 0,
    }
    f = Fault('whatever', my_function, data=dirty_dict)

    dirty_dict['foo'] = 1
    g = Fault('whatever', my_function, data=dirty_dict)

    print(f'Will print False: {f == g}')
    print(f'Will print same string twice, which is wrong: {str(f)}, {str(g)}')

Let's discuss whether we should do deepcopy kwargs or not.