operations/upload: Preserve owner/group/mode of target if exists
As described in the bundlewrap documentation:

 > owner
 >
 > Username of the file's owner. Defaults to 'root'. Set to
 > None if you don't want BundleWrap to change whatever
 > is set on the node.

Setting owner to `None` and uploading a file causes the target to be
owned by the user bundlewrap uses to establish the ssh session, and
overwrites the previous owner of the target. The same principle applies
to group and mode.

`cp` by default does preserves the target permissions if it exists,
(which corresponds to the command line option --no-preserve=all, since
`cp`'s preservation applies to the source file).

Note on `cp -f`: GNU `cp` preserves ownership of the target even using
`cp -f`, but e.g. FreeBSD does not, hence why we use `cp` instead of
`cp -f`.