secrets: Put _load_keys() in "with io.job()"
When you use "key_command = ..." to load the keys in .secrets.cfg, it
can take a while until that command has executed. For example, pass[0]
needs to launch GnuPG to decrypt the files, which might even ask the
user for a passphrase first. During that time, bw just shows an
awkwardly mysterious empty line. It's probably better if we tell users
what's going on.

[0]: https://www.passwordstore.org/