deps: Optimize _prepare_deps()
Cache results of resolve_selector().

This can't be easily done in a generic way with a decorator, because
"items" is a list and not hashable. Besides, the cache would lose its
effect if "originating_item_id" would be taken into account: I'm seeing
lots of queries to, say, "bundle:foo", but all of them have a different
"originating_item_id". So, it's better to just resolve "bundle:foo",
cache the result, and then remove the offending item afterwareds.