items/files: show url instead of nodename+path when downloading file
Showing the node name and the file path seemed to be the correct thing
initially, however i've found that showing that info is often not what
users expect.

Showing the actual url that's being downloaded is much more intuitive
to users. Especially inexperienced bw users have contacted me and asked
why bundlewrap was downloading a file from the node when it should be
*uploading* that file instead.