items/pkg_apt: mark packages as 'explicitely installed'
This fixes #584.

Without marking a package as explicitely installed, a package might come
in as a dependency of another package without bundlewrap noticing it.
This would allow a mechanism such as `apt autoremove` to remove the package,
even though we explicitely want this package to be installed, leading to
possible downtime.